+++ src/components/SVGLabels.js
@@ -1,6 +1,7 @@
 import PureComponent from "./PureComponent";
 import PropTypes from 'prop-types';
 import React from 'react'
+import {HeaderLabel} from "../components/HeaderLabel";
 
 
 let styles = {
@@ -21,31 +22,145 @@ export default class SVGLabels extends PureComponent {
 
     constructor(props){
         super(props);
-        const { width, height ,drawOverlay,...drawProps} = this.props;
+        console.log('SVGLabels input props');
+        console.log(this.props);
+        const { width, height ,...drawProps} = this.props;
+        console.log(drawProps);
 
         this.state = {
             width: width,
             height: height,
-            drawOverlay: drawOverlay,
             drawProps: drawProps,
         }
     }
 
+    drawOverviewLabels(width, height, layout, pathways, selectedPathways, labelHeight, labelOffset, colorMask) {
+
+        if (layout[0].size <= 1) {
+            return;
+        }
+
+        const highestScore = pathways.reduce((max, current) => (max > current.density) ? max : current.density,0);
+
+        if (pathways.length === layout.length) {
+            return layout.map((el, i) => {
+                let d = pathways[i];
+
+                // let color = Math.round(maxColor * (1.0 - (d.density / highestScore)));
+                // let colorString = 'rgb(256,' + color + ',' + color + ')'; // sets the color to fill in the rectangle with
+                let geneLength = d.gene.length;
+                let labelString;
+                if (geneLength === 1) {
+                    labelString = d.gene[0];
+                }
+                else {
+                    labelString = '(' + d.gene.length + ') ';
+                    // pad for 1000, so 4 + 2 parans
+                    while (labelString.length < 5) {
+                        labelString += ' ';
+                    }
+
+                    labelString += d.golabel;
+                }
+                return (
+                    <HeaderLabel
+                        labelHeight={labelHeight}
+                        labelOffset={labelOffset}
+                        highScore={highestScore}
+                        left={el.start}
+                        width={el.size}
+                        item={d}
+                        selectedPathways={selectedPathways}
+                        labelString={labelString}
+                        colorMask={colorMask}
+                        // onMouseClick={selectPathway}
+                        key={labelString}
+                    />
+                )
+            });
+        }
+    }
+
+    drawTissueOverlay(div, props) {
+        let {width, height, layout, referenceLayout, associateData, data: {selectedPathways, pathways, referencePathways}} = props;
+
+        console.log('input data.pathways');
+        console.log(pathways);
+
+
+        if (associateData.length === 0) {
+            alert('Clicked on an empty cell?');
+            return;
+        }
+
+
+        let labels;
+        if (referencePathways) {
+
+            // TODO: for each gene, map the other pathways that gene is involved in
+            let pathwayMapping = pathways.map( p => {
+                return {
+                    label:p.gene[0],
+                    density: p.density,
+                }
+            });
+
+
+            // calculates the scores for the pathways based on existing gene density
+            let newRefPathways = referencePathways.map( r => {
+
+                let density = 0 ;
+
+                // TODO: this could be a lot faster
+                for( let gene of  r.gene){
+                    let found = pathwayMapping.filter( pm => pm.label === gene );
+                    if(found.length>0){
+                        density += found.reduce( (sum , a ) => sum + a.density , 0);
+                    }
+                }
+
+                // TODO: there is a race condition in here, that is messing this up
+                return  {
+                    goid: r.goid,
+                    golabel: r.golabel,
+                    gene: r.gene,
+                    density: density,
+                };
+            });
+
+            console.log('newRefPathways')
+            console.log(newRefPathways)
+
+            console.log('pathways')
+            console.log(pathways)
+
+            let l1 = this.drawOverviewLabels(width, height, referenceLayout, newRefPathways, selectedPathways, 150, 0, [0, 1, 1]);
+            let l2 = this.drawOverviewLabels(width, height, layout, pathways, [], 150, 150, [1, 0, 0]);
+            labels = [...l1, ...l2];
+        }
+        else {
+            labels = this.drawOverviewLabels(width, height, layout, pathways, selectedPathways, 150, 0, [0, 1, 1]);
+        }
+        return labels;
+
+    }
+
     render() {
-        const { width, height ,drawOverlay,onClick,onMouseMove} = this.props;
+        const { width, height ,onClick,onMouseMove} = this.props;
+        console.log('draw props: '+this.state.drawProps)
+        console.log(this.state.drawProps)
         return (
             <div id="expressionOverlay"
                  style={{...styles.overlay, width, height}}
                  onMouseMove={onMouseMove}
                  onClick={onClick}
             >
-                { drawOverlay(this,this.state.drawProps) }
+                { this.drawTissueOverlay(this,this.state.drawProps) }
             </div>
         )
     }
 }
 SVGLabels.propTypes = {
-    drawOverlay: PropTypes.any,
     width: PropTypes.any,
     height: PropTypes.any,
     // onMouseOver: PropTypes.any,
